Rename alloc.rs to allocator.rs
This disambiguates from the libstd alloc module and is also in keeping
with collector.rs.
